% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{calcPK}
\alias{calcPK}
\title{calcPK}
\usage{
calcPK(veep, court, veto, controller, conservative, year = 2018)
}
\value{
a vector with the shape and rate
}
\description{
Calculate political knowledge
}
