% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsSupportFunctions.R
\name{combine.samples}
\alias{combine.samples}
\title{combine.samples}
\usage{
combine.samples(sample1, sample2)
}
\arguments{
\item{sample1}{An mcmc object}

\item{sample2}{Another mcmc object}
}
\value{
The two mcmc objects smushed together
}
\description{
Takes two mcmc objects, and combines the samples together.
}
\examples{
combine.samples(sample1, sample2) #pretty straightforward
}
