% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createChoropleth}
\alias{createChoropleth}
\title{Create World Maps of Bias Data}
\usage{
createChoropleth(data)
}
\arguments{
\item{data}{A dataframe containing the fields: country, biasQC}
}
\description{
This function lets you create plots of the bias across countries on a per-question basis
}
\examples{
createChoropleth()
}
\keyword{proportionJudgments}
