% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createSubjectFileFromQualtricsDemographicData}
\alias{createSubjectFileFromQualtricsDemographicData}
\title{Recode standard political demographic labels and return a subject file}
\usage{
createSubjectFileFromQualtricsDemographicData(data, dataProcessed = "none")
}
\arguments{
\item{data}{a tibble containing columns pk1-pk5, ideology, gender, and education}
}
\description{
This function takes a tibbl containing 
political demographic data, and calculated pk, gender, ideology, etc.
and returns a subject-file with the standard calculated subcomponents.
}
\examples{
createSubjectFileFromQualtricsDemographicData(data)
}
\keyword{TIM}
\keyword{proportionJudgments,}
