% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographicModeling-package.r
\docType{package}
\name{demographicModeling}
\alias{demographicModeling}
\alias{demographicModeling-package}
\title{demographicModeling.}
\description{
This package provides a number of tools we have found useful in modeling demographic estimations, and other kinds of estimations.
This includes one set of functions to calculate, given tibble column names, Della-Carpini & Keeter's political knowledge
assessment, the BNT (coming soon), to recode on political ideology (coming soon), and to calculate and recode 
other standard measures.
}
\details{
It includes a set of functions to build and regularize graphs, and third set of functions to simplify running JAGS models 
over the kinds of data we typically collect from qualtrics, ipsosMORI, and GSS.

It includes a few other convenience functions (such as gammaFromModeSD) that probably belong elsewhere.

To apply the psychophysical functions we typically apply to these kinds of situations, install dlandy/classicPsychophysics

To see a full list of functions, do ls("package:demographicModeling")
}
