% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{gammaFromModeSD}
\alias{gammaFromModeSD}
\title{gammaFromModeSD}
\usage{
gammaFromModeSD(mode = 0.05, sd = 10)
}
\arguments{
\item{mode}{the desired mode value}

\item{sd}{the desired sd}
}
\value{
a vector with the shape and rate
}
\description{
This function takes a mode and SD, and generates the corresponding shape and rate parameters that 
produce a gamma with that set of values.
}
\examples{
gammaFromModeSD(1, 1)
}
