% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{logOdds}
\alias{logOdds}
\title{logOdds}
\usage{
logOdds(x, smallValue = 10^-5)
}
\arguments{
\item{x}{a vector of positive values, between 0 and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses}
}
\value{
a vector containing transformed values
}
\description{
This function is a transformation function that takes the log odds
By design, it is robust to the inclusion of the occasional 0,
which it maps to a very small value, which can optionally be specified
It's appropriate for mapping (0, 1):-> (-inf, inf).
}
\examples{
logOdds(1:100/100)
}
