% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{logOddsMean}
\alias{logOddsMean}
\title{Calculate the mean of a set of proportions, where the mean is done in log odds space.}
\usage{
logOddsMean(proportions)
}
\arguments{
\item{proportions}{A vector of proportions}
}
\description{
Utility function to calculate the proportion whose log-odds is the mean of all the items in the input vector.
}
\examples{
logOddsMean()
}
