% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{logOddsT}
\alias{logOddsT}
\title{Calculate a t-test of a set of proportions, where the test is done in log odds space.}
\usage{
logOddsT(x, y)
}
\arguments{
\item{proportions}{A vector of proportions}
}
\description{
Utility function to calculate the value of a within-t-test over the log-odds of the values.
}
\examples{
logOddsMean()
}
