% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsSupportFunctions.R
\name{par.trace.samples}
\alias{par.trace.samples}
\title{par.trace.samples}
\usage{
par.trace.samples(file, data, n.adapt = 500, inits = NULL, n.iter = 1000,
  n.samples = 500, thin = 1, monitor, n.breaks = c(10, 2, 5),
  predictionVariable = FALSE)
}
\arguments{
\item{file}{A jags file}

\item{data}{Any data values, in a list, as though you were passing them to jags.model (you are)}

\item{n.adapt}{The number of adapts to do}

\item{inits}{Initial parameters}

\item{n.iter}{A number of iterations to run on the update cycle}

\item{n.samples}{The number of samples to collect.}

\item{thin}{How much to thin}

\item{monitor}{What variables do we want to monitor?}

\item{n.breaks}{How often should we pause to update the user about progress?}

\item{predictionVariable}{If there is a prediction variable, that variable can be produced in a special cycle, to avoid generating too much data}
}
\value{
a list of samples
}
\description{
Composite function that handles creating, initing, and running jags models in parallel, then collects
the resulting samples.
}
\examples{
results <- par.trace.samples(file=paste(workingDir, jagsFile , sep="")
    , data=jagsData
    , monitor=monitor
    , n.iter = updateNum
    , n.adapt = n.adapt
    , n.samples = nSamplesCoda
    , n.breaks=5
    , thin=thinning
    )
}
