% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{printQuestionParameterPDF}
\alias{printQuestionParameterPDF}
\title{Create PDF's of parameters}
\usage{
printQuestionParameterPDF(graphs, file = "parameters.pdf", items = c(1:3))
}
\arguments{
\item{graphs}{the output of a call to createQuestionLevelProportionJudgmentPlots}

\item{file}{the name of the pdf file to use}

\item{items}{3 items, of those returned by createQuestionLevelProportionJudgmentPlots, to plot}
}
\description{
This function creates a pdf of your data. You must give it the output of a 
call to createQuestionLevelProportionJudgmentPlots. and a set of 3 items
}
\examples{
printQuestionParameterPDF()
}
\keyword{proportionJudgments}
