% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{sErrors}
\alias{sErrors}
\title{sErrors}
\usage{
sErrors(x)
}
\arguments{
\item{x}{a vector of  values,}
}
\value{
the standard error
}
\description{
This function is a transformation function that takes the log odds
By design, it is robust to the inclusion of the occasional 0,
which it maps to a very small value, which can optionally be specified
It's appropriate for mapping (0, 1):-> (-inf, inf).
}
\examples{
sErrors(rnorm(100))
}
