% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createQuestionLevelProportionJudgmentPlots}
\alias{createQuestionLevelProportionJudgmentPlots}
\title{Create Question-Level Proportion Judgment Plots}
\usage{
createQuestionLevelProportionJudgmentPlots(data, year = "",
  countryVariabilityBars = FALSE)
}
\arguments{
\item{data}{A dataframe containing the fields: question, bias, biasQCTau, biasLow, biasHigh, and ditto for lambda and tau}
}
\description{
This function lets you create plots of the bias, ignorance, and cross-unit variability on a per-question basis,
From results of a JAGS model.  If you want to print those on a common structure, feed the output into printQuestionParameterPDF
}
\examples{
createQuestionLevelProportionJudgmentPlots()
}
\seealso{
printQuestionParameterPDF
}
\keyword{proportionJudgments}
