% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{sErrors}
\alias{sErrors}
\title{sErrors}
\usage{
sErrors(x)
}
\arguments{
\item{x}{a vector of  values,}
}
\value{
the standard error
}
\description{
This function takes the sd/sqrt(length)
}
\examples{
sErrors(rnorm(100))
}
